refactor: modern_settings_dialog 控件库与样式表剥离（批6-7）
- 12 个自定义控件类 → settings_widgets.py（AST 逐字符保真）
- 三段内联 QSS → settings_styles*.qss 文件（逐字节一致）
- 三平台构建脚本补新 qss 的 --add-data（全变体无条件段；
  缺失会导致打包后设置页样式丢失）
- 5.6sol 盲审：2 个 P3 均为理论性（无存活消费方），记录接受

结构优化批6-7（STRUCT_PLAN_FINAL.md）
