+++ pet/settings_styles.qss
@@ -1,0 +1,122 @@
+
+QDialog {
+    background: #fcfcfd;
+    color: #202020;
+    font-family: "SF Pro Text", ".AppleSystemUIFont", "PingFang SC", "Segoe UI", sans-serif;
+    font-size: 13px;
+}
+QFrame#sidebarPane {
+    background: #f7f7f8;
+    border: none;
+    border-right: 1px solid #e3e5e8;
+}
+QStackedWidget { background: #fcfcfd; }
+QLineEdit#settingsSearch {
+    min-height: 30px;
+    padding: 0 8px;
+    background: #f0f1f3;
+    border: 1px solid transparent;
+    border-radius: 15px;
+    color: #202020;
+}
+QLineEdit#settingsSearch:focus {
+    border: 2px solid #0a84ff;
+    padding: 0 7px;
+}
+QPushButton#saveAndExit {
+    min-height: 28px;
+    padding: 2px 8px;
+    text-align: left;
+    background: transparent;
+    border: none;
+    border-radius: 8px;
+    font-weight: 500;
+}
+QPushButton#saveAndExit:hover { background: #e9eaec; }
+QLabel#searchStatus {
+    padding: 0 5px;
+    color: #777b80;
+    font-size: 11px;
+}
+QListWidget#settingsSidebar {
+    background: transparent;
+    border: none;
+    outline: none;
+    font-size: 13px;
+    font-weight: 500;
+}
+QListWidget#settingsSidebar::item {
+    min-height: 26px;
+    padding: 4px 10px;
+    border-radius: 9px;
+    color: #4e4e4e;
+}
+QListWidget#settingsSidebar::item:hover {
+    background: #eceef1;
+    color: #202020;
+}
+QListWidget#settingsSidebar::item:selected {
+    background: #e3e5e8;
+    color: #171717;
+}
+QLabel#pageTitle {
+    font-size: 26px;
+    font-weight: 600;
+    color: #171717;
+}
+QLabel#sectionTitle {
+    font-size: 14px;
+    font-weight: 600;
+    color: #2b2b2b;
+}
+QFrame#settingsCard {
+    background: #ffffff;
+    border: 1px solid #e2e4e8;
+    border-radius: 11px;
+}
+QFrame#cardSeparator {
+    background: #eceef1;
+    border: none;
+    margin-left: 14px;
+    margin-right: 14px;
+}
+QLabel#settingLabel {
+    font-size: 14px;
+    font-weight: 600;
+    color: #252525;
+}
+QLabel#settingHint {
+    font-size: 12px;
+    font-weight: 400;
+    color: #777777;
+}
+QLabel#settingLabel:disabled, QLabel#settingHint:disabled { color: #a6a8ac; }
+SettingRow[searchMatch="true"] {
+    background: #eaf3ff;
+    border-radius: 8px;
+}
+QScrollArea#settingsScroll, QScrollArea#settingsScroll > QWidget > QWidget {
+    background: transparent;
+}
+QListWidget#quickLaunchList {
+    background: #fbfbfb;
+    border: 1px solid #d9d9d9;
+    border-radius: 8px;
+    outline: none;
+    padding: 3px;
+}
+QListWidget#quickLaunchList::item {
+    min-height: 30px;
+    padding: 3px 7px;
+    border-radius: 6px;
+}
+QListWidget#quickLaunchList::item:selected { background: #e8e8e8; color: #202020; }
+QPushButton {
+    min-height: 26px;
+    padding: 1px 12px;
+    background: #ffffff;
+    border: 1px solid #d0d0d0;
+    border-radius: 7px;
+    font-weight: 500;
+}
+QPushButton:hover { background: #f0f0f0; }
